Fixed progress bar not visible when editing remote files; moved edit and tricky_copy to actions; added new function recv/send one
